% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guidelines}
\alias{guidelines}
\title{Fabricated guidelines data}
\format{A data frame with 11 variables:
\describe{
\item{\code{Zone}}{A major spatial unit}
\item{\code{Component}}{The highest level of the variable
    hierarchy}
\item{\code{Indicator}}{Indicator level}
\item{\code{Subindicator}}{Subindicator level}
\item{\code{Measure}}{Measure level.  The lowest level of the
    variable hierarchy}
\item{\code{GL}}{The guideline/threshold values}
\item{\code{Lower, Upper}}{Lower and upper bounds for guidelines that apply
    to a range (for example pH of Dissolved oxygen)}
\item{\code{DOF}}{Direction of Failure.  This will be either 'H'
    for higher (indicating that values higher than the guideline
    consistute an excedence of the guideline), 'L' for lower than
    the guideline or 'B' for a guideline range.}
\item{\code{Cuts}}{Break=Score pairs for defining the cuts in the
     data and corresponding scores.}
}}
\usage{
guidelines
}
\description{
Fabricated guidelines data
}
\seealso{
\code{\link{reportcardData}}
}
\keyword{datasets}
