% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator.hier}
\alias{indicator.hier}
\title{Fabricated indicator hierarchy}
\format{A data frame with 4 variables:
\describe{
\item{\code{Component}}{The highest level of the measure
    hierarchy}
\item{\code{Indicator}}{Indicator level}
\item{\code{Subindicator}}{Subindicator level}
\item{\code{Measure}}{Measure level.  The lowest level of the
    variable hierarchy}
}}
\usage{
indicator.hier
}
\description{
Fabricated indicator hierarchy
}
\seealso{
\code{\link{reportcardData}}, \code{\link{spatial.hier}}, \code{\link{guidelines}}
}
\keyword{datasets}
