% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{RC_aggregate}
\alias{RC_aggregate}
\title{Simple hierarchical aggregation}
\usage{
RC_aggregate(df, grouping_cols = colnames(df)[!colnames(df) \%in\%
  c("Score", "Grade")], fun = weighted.mean, gradetype = "Uniform")
}
\arguments{
\item{df}{a data frame or tibble containing at least the following
fields: Score, Weights}

\item{grouping_cols}{a vector of strings indicating which columns
to group the aggregation by}

\item{fun}{a function used in the aggregation}

\item{type}{a string indicating the conversion method.  Must be one
of 'Uniform' (for Uniform grade boundaries used by the GBR
Report Card), 'MMP' (for grade boundaries used by the GBR
Marine Monitoring Program Report Card), GHHP (for grade
boundaries used by the Gladstone Healthy Harbour Partnership
Report Card) or 'MCWC' (for the grade boundaries used by the
MidCoast Council Waterway and Catchment Report Card).}
}
\value{
an aggregated data frame (tibble)
}
\description{
Simple hierarchical aggregation
}
\examples{
library(tidyverse)
data(reportCard)
reportCard.idx = reportCard \%>\%      #generate fsMAMP indices
    mutate(Score=RC_index(x=Value, GL=GL, Lower=Lower, #calculate index
    Upper=Upper, fold=2, DOF=DOF, type='MAMP', capped=TRUE,
    scaled=TRUE))
## Aggregate over all the samples collected throughout the year
reportyear.site.measure = reportCard.idx \%>\%
    mutate(Weight=1) \%>\%   # add weights
    RC_aggregate(grouping_cols=c('ReportYear','Zone','Site','Component','Indicator','Subindicator', 'Measure'))
head(reportyear.site.measure)
}
\seealso{
\code{\link{RC_boot_aggregate}}, \code{\link{RC_index}}
}
\author{
Murray Logan
}
