% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{RC_boot_accumulate}
\alias{RC_boot_accumulate}
\title{Bootstrapp accumulation}
\usage{
RC_boot_accumulate(boot = NULL, size = 10, seed = 123,
  grouping_cols = colnames(df)[!colnames(df) \%in\% c("Score", "Grade")])
}
\arguments{
\item{boot}{a data frame or tibble containing at least the
following fields: Score, Weights}

\item{size}{an integer, the number of resamples to generate the
accumulated distributions}

\item{seed}{an integer, the random seed to apply}

\item{grouping_cols}{a vector of strings indicating which columns
to group the aggregation by.}
}
\value{
a list of length 2:
    dist: the bootstrapp samples
    sum:  summarized
}
\description{
This function performs bootstrapp accumulation.  The
    accumulation groups together single observations into a
    distribution (needed for meaningful bootstrapp aggregation)
}
\details{
Put the details in
}
\examples{
library(tidyverse)
data(reportCard)
reportCard.idx = reportCard \%>\%      #generate fsMAMP indices
    mutate(Score=RC_index(x=Value, GL=GL, Lower=Lower, #calculate index
    Upper=Upper, fold=2, DOF=DOF, type='MAMP', capped=TRUE,
    scaled=TRUE))
## Accummulate over all samples collected throught each year
seed=123
size=100
reportyear.site.measure.boot=reportCard.idx \%>\%
    mutate(Weight=1) \%>\%   # add weights
    RC_boot_accumulate(size=size, seed=seed, grouping_cols=c('ReportYear','Zone','Site','Component','Indicator','Subindicator', 'Measure'))
reportyear.site.measure.boot$sum
}
\seealso{
\code{\link{RC_boot_aggregate}}, \code{\link{RC_aggregate}}
}
\author{
Murray Logan
}
