% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{RC_boot_aggregate}
\alias{RC_boot_aggregate}
\title{Bootstrapp aggregation of Report Card Scores}
\usage{
RC_boot_aggregate(boot = NULL, size = 10, seed = 123,
  grouping_cols = colnames(df)[!colnames(df) \%in\% c("Score", "Grade")],
  over = "", fun = weighted.mean)
}
\arguments{
\item{boot}{a data frame or tibble containing at least the
following fields: Score, Weights}

\item{size}{an integer, the number of bootstrapp samples}

\item{seed}{an integer, the random seed to apply}

\item{grouping_cols}{a vector of strings indicating which columns
to group the aggregation by.}

\item{over}{a string indicating the column (categorical variable)
to aggregate over}

\item{fun}{the function to apply in the aggregation}
}
\value{
a list of length 2:
    dist: the bootstrapp samples
    sum:  summarized
}
\description{
This function performs bootstrapp aggregation.
}
\details{
Put the details in
}
\examples{
library(tidyverse)
data(reportCard)
reportCard.idx = reportCard \%>\%      #generate fsMAMP indices
    mutate(Score=RC_index(x=Value, GL=GL, Lower=Lower, #calculate index
    Upper=Upper, fold=2, DOF=DOF, type='MAMP', capped=TRUE,
    scaled=TRUE))
## Accummulate over all samples collected throught each year
seed=123
size=10
reportyear.site.measure.boot=reportCard.idx \%>\%
    mutate(Weight=1) \%>\%   # add weights
    RC_boot_accumulate(size=size, seed=seed, grouping_cols=c('ReportYear','Zone','Site','Component','Indicator','Subindicator', 'Measure'))
## Aggregate over Measures within ReportYear/Site/Subindicators
reportyear.site.subindicator.boot = reportyear.site.measure.boot$dist \%>\%
    mutate(Weight=1) \%>\%   # add weights
    RC_boot_aggregate(size=size, seed=seed,
    grouping_cols=c('ReportYear','Zone','Site','Component','Indicator','Subindicator'), over='Measure')
reportyear.site.subindicator.boot$sum
}
\seealso{
\code{\link{RC_boot_accumulate}},
    \code{\link{RC_boot_aggregate}}
}
\author{
Murray Logan
}
