% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_cuts_matrix}
\alias{RC_cuts_matrix}
\title{Convert cut string into matrix}
\usage{
RC_cuts_matrix(x)
}
\arguments{
\item{x}{a string containing the ':' separated Break=Score cut pairs}
}
\value{
a matrix of Breaks and corresponding Scores
}
\description{
Convert a string of cuts (Break=Score pairs separated
    by ':') into a matrix in which each row represents a
    Break/Score pair
}
\author{
Murray Logan
}
