% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_financialYear}
\alias{RC_financialYear}
\title{Determine financial year}
\usage{
RC_financialYear(Date)
}
\arguments{
\item{Date}{a vector of Dates}
}
\value{
a vector of integers indicating the financial year
}
\description{
This function uses a vector of valid \code{Dates} to
    create a new vector that indicates the corresponding financial
    year.
}
\examples{
RC_financialYear(.leap.seconds)
}
\seealso{
\code{\link{RC_waterYear}}
}
