% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{RC_generateGrades}
\alias{RC_generateGrades}
\title{Generate Grades}
\usage{
RC_generateGrades(x, type = "MMP")
}
\arguments{
\item{x}{a vector of numerical values representing the report card
Scores.  These values should be in the range of 0-1.}

\item{type}{a string indicating the conversion method.  Must be one
of 'Uniform' (for Uniform grade boundaries used by the GBR
Report Card), 'MMP' (for grade boundaries used by the GBR
Marine Monitoring Program Report Card), GHHP (for grade
boundaries used by the Gladstone Healthy Harbour Partnership
Report Card) or 'MCWC' (for the grade boundaries used by the
MidCoast Council Waterway and Catchment Report Card).}
}
\value{
a vector of alpha-numeric Grades ('A', 'B', 'C', 'D' or 'E').
}
\description{
Convert report card Scores into Grades.
}
\examples{
RC_generateGrades(x=seq(0,1,length=6), type='Uniform')
}
