% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{RC_index}
\alias{RC_index}
\title{Index formulations}
\usage{
RC_index(x = NA, GL = NA, Lower = NA, Upper = NA, lBound = NA,
  uBound = NA, Cuts = NA, fold = 2, DOF = "H", T = NA,
  type = "Binary", capped = FALSE, scaled = FALSE)
}
\arguments{
\item{x}{a numeric vector of observed values}

\item{GL}{a numeric vector the same length as \code{x}
representing the Guideline/reference/threshold value}

\item{Lower}{a numeric vector the same length as \code{x}
representing the lower range for items that have an ideal
optimum.  Only applied when DOF='B'.}

\item{Upper}{a numeric vector the same length as \code{x}
representing the upper range for items that have an ideal
optimum.  Only applied when DOF='B'.}

\item{lBound}{a numeric vector the same length as \code{x}
representing the lower bound of the index scaling and capping
region.  If non NA, \code{fold} will be ignored.}

\item{uBound}{a numeric vector the same length as \code{x}
representing the upper bound of the index scaling and capping
region. If non NA, \code{fold} will be ignored.}

\item{Cuts}{a character vector indicating, the conversion chart
breaks and corresponding scores (via key=value pairs
separated by ':')}

\item{fold}{a numeric vector the same length as \code{x}.  Applied
when \code{type='MAMP'}, indicating the multiplier and
divisory applied to the \code{GL} (on a log_2 scale) for
determining the lower and upper bounds of the index scaling
and capping region.}

\item{DOF}{a character vector the same length as \code{x}
representing an indication of the Direction Of Failure.  Must
be one of 'H' (higher), 'L' (lower) or 'B' (both).}

\item{type}{an abbreviation of the index formulation to use.  Must
be one of 'Binary' (Binary), 'MAMP' (Modified Amplitude).}

\item{capped}{a logical indicating whether the indices should be
capped to the range [\code{lBound}, \code{uBound}].}

\item{scaled}{a logical indicating whether the indices should be
scaled to the range [0,1].  This cannot be applied to HSAMP or
LAMP, the former of which is already scaled.}
}
\value{
a numeric vector of index values
}
\description{
Index formulations including 'Binary'.
}
\details{
Binary formulation:
\deqn{Score_i = \left\{\begin{array}{l l}1 & \text{if}~x_i \le
    G_i\\0 & \text{if}~x_i~\text{else}\\\end{array}}{
Score_i = [if x_i <= GL_i] 1
          [else:] 0
}
Benchmark and Worst Case Scenario formulation:
Should not be applied to items with an optimum.
Modified Amplitude formulation:
\deqn{}{
Score_i = [if x_i > GL_i = fail:]
                 log_2(x_i/G_i)^-1\cr
          [if x_i < GL = fail:]
                 log_2(x_i/G_i)^1
}
}
\examples{
library(tidyverse)
## A very simple example
RC_index(x=c(30,40,60,70), GL=50, DOF='H', type='Binary')

## Using built-in fabricated ecological monitoring and guidelines
##    data
library(tidyverse)
data(reportCard)
reportCard.idx = reportCard \%>\%
mutate(Score = case_when(
              Measure \%in\% c('Chla') ~ RC_index(x = Value, GL = GL, fold = 2, DOF = DOF, type = 'MAMP', capped = TRUE, scaled = TRUE),
              Measure \%in\% c('TN', 'TP') ~ RC_index(x = Value, GL = GL, lBound = lBound, uBound = uBound, T = 2, DOF = DOF, type = 'HSAMP'),
              Measure \%in\% c('Biomass', 'Secchi') ~ RC_index(x = Value, GL = GL, lBound = lBound, uBound = uBound, T = 3, DOF = DOF, type = 'HSAMP'),
              Measure \%in\% c('pH') ~ RC_index(x = Value, Lower = Lower, Upper = Upper, lBound = lBound, uBound = uBound, T = 3, DOF = DOF, type = 'HSAMP'),
              Measure \%in\% c('Cover') ~ RC_index(x = Value, Cuts = Cuts, type = 'Chart')
      )
   )
reportCard.idx \%>\% as.data.frame \%>\% head
}
\seealso{
\code{\link{RC_visualize_indices}}
}
\author{
Murray Logan
}
