% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_makeWetDrySeason}
\alias{RC_makeWetDrySeason}
\title{Determine Austral tropical Season (Wet/Dry)}
\usage{
RC_makeWetDrySeason(Date)
}
\arguments{
\item{Date}{a vector of Dates}
}
\value{
a vector of strings indicating either 'Wet' or 'Dry'
}
\description{
This function uses a vector of valid \code{Dates} to
    create a new vector that indicates the corresponding Austral
    Tropical Season (Wet or Dry).
}
\details{
The Austral 'Wet' Season is defined as the months of November,
    December, January, February, March and April.  The Austral
    'Dry' Season is defined as the months of May, June, July,
    August, September and October.
}
\examples{
RC_makeWetDrySeason(.leap.seconds)
}
\seealso{
\code{\link{RC_waterYear}}
}
