% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_max_cut}
\alias{RC_max_cut}
\title{Extract the Break associated with the maximum Score}
\usage{
RC_max_cut(x)
}
\arguments{
\item{x}{a string containing the ':' separated Break=Score cut pairs}
}
\value{
a numeric representing the Break associated with the
    maximum Score
}
\description{
Extract the Break associated with the maximum Score
    from a sting of ':' separated Break=Score cut pairs
}
\author{
Murray Logan
}
