% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_min_cut}
\alias{RC_min_cut}
\title{Extract the Break associated with the minimum Score}
\usage{
RC_min_cut(x)
}
\arguments{
\item{x}{a string containing the ':' separated Break=Score cut pairs}
}
\value{
a numeric representing the Break associated with the
    minimum Score
}
\description{
Extract the Break associated with the minimum Score
    from a sting of ':' separated Break=Score cut pairs
}
\author{
Murray Logan
}
