% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_reportCardColorPalette}
\alias{RC_reportCardColorPalette}
\title{Report Card Traffic light colour palette}
\usage{
RC_reportCardColorPalette(n)
}
\value{
a color palette
}
\description{
A colour palette associated with the
    report card traffic light colour scheme
}
\author{
Murray Logan
}
