% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{RC_reportCardColors}
\alias{RC_reportCardColors}
\title{Report Card Traffic light colours}
\format{An object of class \code{character} of length 5.}
\usage{
RC_reportCardColors
}
\value{
a vector of hex codes
}
\description{
A vector of hex colour codes associated with the
    report card traffic light colour scheme
}
\author{
Murray Logan
}
\keyword{datasets}
