% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{RC_visualize_indices}
\alias{RC_visualize_indices}
\title{Visualize indices}
\usage{
RC_visualize_indices(type = NA, fold = 2, T = 2, capped = FALSE,
  scaled = FALSE, min = NA, max = NA, GL = 50, Lower = 40,
  Upper = 60, lBound = 20, uBound = 80,
  cutsH = "0=0:15=0.2:30=0.4:50=0.6:65=0.8:80=1:100=1",
  cutsL = "0=1:5=1:15=0.8:30=0.6:50=0.4:65=0.2:80=0:100=0",
  cutsB = "0=0:10=0.20:20=0.35:30=0.5:40=0.7:45=0.9:55=0.9:60=0.7:70=0.5:80=0.35:90=0.2:100=0")
}
\arguments{
\item{type}{an abbreviation of the index formulation to use.  Must
be one of 'Binary' (Binary), 'MAMP' (Modified Amplitude).}

\item{fold}{a numeric vector the same length as \code{x}.  Applied
when \code{type='MAMP'}, indicating the multiplier and
divisory applied to the \code{GL} (on a log_2 scale) for
determining the lower and upper bounds of the index scaling
and capping region.}

\item{capped}{a logical indicating whether the indices should be
capped to the range [\code{lBound}, \code{uBound}].}

\item{scaled}{a logical indicating whether the indices should be
scaled to the range [0,1].  This cannot be applied to HSAMP or
LAMP, the former of which is already scaled.}

\item{min, }{max minimum and maximum expected data range}

\item{GL}{guideline value}

\item{Lower, }{Upper lower and upper guideline values (when
guidelines are a range)}

\item{lBound, }{uBound lower and upper scaling and capping limits}

\item{cutsH, }{cutsL, cutsB a string of Break:Score pairs for Chart
index formulation.}
}
\value{
a ggplot
}
\description{
Visualize the general response curve associated with a
    index formulation.
}
\examples{
library(tidyverse)
RC_visualize_indices(type='WCS', scaled=TRUE)
RC_visualize_indices(type='WCS', scaled=TRUE)
RC_visualize_indices(type='MAMP', fold=2)
RC_visualize_indices(type='MAMP', fold=2, scaled=TRUE)
RC_visualize_indices(type='MAMP', fold=2, capped=TRUE, scaled=FALSE)
RC_visualize_indices(type='MAMP', fold=5, capped=TRUE, scaled=FALSE)
RC_visualize_indices(type='MAMP', fold=5, capped=TRUE, scaled=TRUE)
RC_visualize_indices(type='MAMP', fold=2, capped=TRUE, scaled=TRUE)
RC_visualize_indices(type='HSAMP')
RC_visualize_indices(type='HSAMP', T=5)
RC_visualize_indices(type='LAMP', T=2)
RC_visualize_indices(type='LAMP', T=7)
RC_visualize_indices(type='Chart')
}
\author{
Murray Logan
}
