% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RC_waterYear}
\alias{RC_waterYear}
\title{Determine the Austral tropical water year}
\usage{
RC_waterYear(Date)
}
\arguments{
\item{Date}{a vector of Dates}
}
\value{
a vector of integers indicating the water year
}
\description{
This function uses a vector of valid \code{Dates} to
    create a new vector that indicates the corresponding Austral
    Tropical water year
}
\details{
The Austral Tropical water year is defined as between October
    1st and September 30.
}
\examples{
RC_waterYear(.leap.seconds)
}
\seealso{
\code{\link{RC_financialYear}}
}
