% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reportCard}
\alias{reportCard}
\title{Fabricated report card data and guidelines (combined)}
\format{A data frame with 19 variables:
\describe{
\item{\code{Date}}{Sample date}
\item{\code{Site}}{A lower level spatial unit nested within Zone}
\item{\code{Measure}}{The measured/observed variable}
\item{\code{Value}}{The observed value}
\item{\code{Month}}{The sampling month}
\item{\code{ReportYear}}{Reporting year base on an Austral
    Tropical Water year (1st Oct - 30th Sept)}
\item{\code{Region}}{The top level of the spatial hierarchy}
\item{\code{Zone}}{A spatial unit nested within \code{Region}}
\item{\code{Component}}{The highest level of the variable
    hierarchy}
\item{\code{Indicator}}{Indicator level}
\item{\code{Subindicator}}{Subindicator level}
\item{\code{GL}}{The guideline/threshold value}
\item{\code{Lower, Upper}}{Lower and upper bounds for guidelines that apply
    to a range (for example pH of Dissolved oxygen)}
\item{\code{lBound, uBound}}{Lower and upper capping and scaling bounds}
\item{\code{DOF}}{Direction of Failure.  This will be either 'H'
    for higher (indicating that values higher than the guideline
    consistute an excedence of the guideline), 'L' for lower than
    the guideline or 'B' for a guideline range.}
\item{\code{Cuts}}{Break=Score pairs for defining the cuts in the
    data and corresponding scores.}
}}
\usage{
reportCard
}
\description{
Fabricated report card data and guidelines (combined)
}
\seealso{
\code{\link{reportcardData}} \code{\link{indicator.hier}}, \code{\link{spatial.hier}}, \code{\link{guidelines}}
}
\keyword{datasets}
