% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reportcardData}
\alias{reportcardData}
\title{Fabricated report card data}
\format{A data frame with 9 variables:
\describe{
\item{\code{Date}}{Sample date}
\item{\code{Site}}{A lower level spatial unit nested within Zone}
\item{\code{Biomass}}{Seagrass Biomass}
\item{\code{Chl-a}}{Chlorophyl-a concentration}
\item{\code{Cover}}{Percent coral cover}
\item{\code{pH}}{pH}
\item{\code{Secchi}}{Secchi depth}
\item{\code{TN}}{Total nitrogen concentration}
\item{\code{TP}}{Total phosphorus concentration}
}}
\usage{
reportcardData
}
\description{
Fabricated report card data
}
\seealso{
\code{\link{reportCard}}, \code{\link{indicator.hier}}, \code{\link{spatial.hier}}, \code{\link{guidelines}}
}
\keyword{datasets}
