% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial.hier}
\alias{spatial.hier}
\title{Fabricated spatial hierarchy}
\format{A data frame with 3 variables:
\describe{
\item{\code{Region}}{The top level of the spatial hierarchy}
\item{\code{Zone}}{a spatial unit nested within \code{Region}}
\item{\code{Site}}{the fixed locations at which the measurements
    were collected. These are nested within \code{Zone}}
}}
\usage{
spatial.hier
}
\description{
Fabricated spatial hierarchy
}
\seealso{
\code{\link{reportcardData}}, \code{\link{indicator.hier}}, \code{\link{guidelines}}
}
\keyword{datasets}
